teste create repository

+++ utils/random.go
@@ -37,6 +37,6 @@ func RandomCategory() string {
 
 // RandomCode generates a random amount of number
 
-func RandomCode() int {
-	return int(RandomInt(0, 1000))
+func RandomCode() int64 {
+	return int64(RandomInt(0, 1000))
 }
